@@ -1,3 +1,5 @@
+// ignore_for_file: use_build_context_synchronously
+
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
@@ -8,9 +10,9 @@ class Login extends StatefulWidget {
   final FirebaseFirestore firestore;
 
   const Login({
-    Key key,
-    @required this.auth,
-    @required this.firestore,
+    required Key key,
+    required this.auth,
+    required this.firestore,
   }) : super(key: key);
   @override
   _LoginState createState() => _LoginState();
@@ -25,69 +27,72 @@ class _LoginState extends State<Login> {
       body: Center(
         child: Padding(
           padding: const EdgeInsets.all(60.0),
-          child: Builder(builder: (BuildContext context) {
-            return Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                TextFormField(
-                  key: const ValueKey("username"),
-                  textAlign: TextAlign.center,
-                  decoration: const InputDecoration(hintText: "Username"),
-                  controller: _emailController,
-                ),
-                TextFormField(
-                  key: const ValueKey("password"),
-                  textAlign: TextAlign.center,
-                  decoration: const InputDecoration(hintText: "Password"),
-                  controller: _passwordController,
-                ),
-                const SizedBox(
-                  height: 20,
-                ),
-                RaisedButton(
-                  key: const ValueKey("signIn"),
-                  onPressed: () async {
-                    final String retVal = await Auth(auth: widget.auth).signIn(
-                      email: _emailController.text,
-                      password: _passwordController.text,
-                    );
-                    if (retVal == "Success") {
-                      _emailController.clear();
-                      _passwordController.clear();
-                    } else {
-                      Scaffold.of(context).showSnackBar(
-                        SnackBar(
-                          content: Text(retVal),
-                        ),
+          child: Builder(
+            builder: (BuildContext context) {
+              return Column(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  TextFormField(
+                    key: const ValueKey("username"),
+                    textAlign: TextAlign.center,
+                    decoration: const InputDecoration(hintText: "Username"),
+                    controller: _emailController,
+                  ),
+                  TextFormField(
+                    key: const ValueKey("password"),
+                    textAlign: TextAlign.center,
+                    decoration: const InputDecoration(hintText: "Password"),
+                    controller: _passwordController,
+                  ),
+                  const SizedBox(
+                    height: 20,
+                  ),
+                  ElevatedButton(
+                    key: const ValueKey("signIn"),
+                    onPressed: () async {
+                      final String? retVal =
+                          await Auth(auth: widget.auth).signIn(
+                        email: _emailController.text,
+                        password: _passwordController.text,
                       );
-                    }
-                  },
-                  child: const Text("Sign In"),
-                ),
-                FlatButton(
-                  key: const ValueKey("createAccount"),
-                  onPressed: () async {
-                    final String retVal =
-                        await Auth(auth: widget.auth).createAccount(
-                      email: _emailController.text,
-                      password: _passwordController.text,
-                    );
-                    if (retVal == "Success") {
-                      _emailController.clear();
-                      _passwordController.clear();
-                    } else {
-                      Scaffold.of(context).showSnackBar(
-                        SnackBar(
-                          content: Text(retVal),
-                        ),
+                      if (retVal == "Success") {
+                        _emailController.clear();
+                        _passwordController.clear();
+                      } else {
+                        Scaffold.of(context).showSnackBar(
+                          SnackBar(
+                            content: Text(retVal!),
+                          ),
+                        );
+                      }
+                    },
+                    child: const Text("Sign In"),
+                  ),
+                  TextButton(
+                    key: const ValueKey("createAccount"),
+                    onPressed: () async {
+                      final String? retVal =
+                          await Auth(auth: widget.auth).createAccount(
+                        email: _emailController.text,
+                        password: _passwordController.text,
                       );
-                    }
-                  },
-                  child: const Text("Create Account"),
-                )
-              ],
-            );
-          }),
+                      if (retVal == "Success") {
+                        _emailController.clear();
+                        _passwordController.clear();
+                      } else {
+                        Scaffold.of(context).showSnackBar(
+                          SnackBar(
+                            content: Text(retVal!),
+                          ),
+                        );
+                      }
+                    },
+                    child: const Text("Create Account"),
+                  ),
+                ],
+              );
+            },
+          ),
         ),
       ),
     );
